+++ client/acp/src/model/setting-fields.js
@@ -1,3 +1,5 @@
 export const ACTIVITY_LIMIT = 'activityLimit';
 
+export const MAX_REWARDS_PER_ACCOUNT = 'maxRewardsPerAccount';
+
 export const MAX_REWARDS_PER_POST = 'maxRewardsPerPost';