+++ client/acp/src/view/widget/settings.js
@@ -18,7 +18,19 @@ class Settings extends React.Component {
         return (
             <div className="settings">
                 <div className="row">
-
+                    <div className="col-md-6">
+                        <div className="form-group">
+                            <label htmlFor="maxRewardsPerAccount">Max Rewards Per Account</label>
+                            <input
+                                className="form-control"
+                                type="text"
+                                id="maxRewardsPerAccount"
+                                onChange={e => this.props.changeField(SettingFields.MAX_REWARDS_PER_ACCOUNT, e.target.value)}
+                                value={this.props.settings[SettingFields.MAX_REWARDS_PER_ACCOUNT]}/>
+                            <p className="help-block">A number of rewards to inject in an account. Specify -1 to remove
+                                a limit, or 0 to disable a feature.</p>
+                        </div>
+                    </div>
                     <div className="col-md-6">
                         <div className="form-group">
                             <label htmlFor="maxRewardsPerPost">Max Rewards Per Post</label>
@@ -31,6 +43,11 @@ class Settings extends React.Component {
                             <p className="help-block">A number of rewards to inject in a post. Specify -1 to remove a
                                 limit, or 0 to disable a feature.</p>
                         </div>
+                    </div>
+                </div>
+
+                <div className="row">
+                    <div className="col-md-6">
                     </div>
                     <div className="col-md-6">
                         <div className="form-group">